Creating FileSystemMasterClient and FileSystemWorkerClient interfaces.
