+++ core/client/src/main/java/alluxio/client/file/FileSystemMasterClient.java
@@ -11,9 +11,8 @@
 
 package alluxio.client.file;
 
-import alluxio.AbstractMasterClient;
 import alluxio.AlluxioURI;
-import alluxio.Constants;
+import alluxio.MasterClient;
 import alluxio.client.file.options.CheckConsistencyOptions;
 import alluxio.client.file.options.CompleteFileOptions;
 import alluxio.client.file.options.CreateDirectoryOptions;
@@ -25,69 +24,14 @@ import alluxio.client.file.options.LoadMetadataOptions;
 import alluxio.client.file.options.MountOptions;
 import alluxio.client.file.options.SetAttributeOptions;
 import alluxio.exception.AlluxioException;
-import alluxio.thrift.AlluxioService;
-import alluxio.thrift.AlluxioTException;
-import alluxio.thrift.FileSystemMasterClientService;
-import alluxio.wire.ThriftUtils;
-
-import org.apache.thrift.TException;
 
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.util.ArrayList;
 import java.util.List;
 
-import javax.annotation.concurrent.ThreadSafe;
-import javax.security.auth.Subject;
-
 /**
- * A wrapper for the thrift client to interact with the file system master, used by alluxio clients.
- *
- * Since thrift clients are not thread safe, this class is a wrapper to provide thread safety, and
- * to provide retries.
+ * A client to use for interacting with a file system master.
  */
-@ThreadSafe
-public final class FileSystemMasterClient extends AbstractMasterClient {
-  private FileSystemMasterClientService.Client mClient = null;
-
-  /**
-   * Creates a new file system master client.
-   *
-   * @param subject the subject
-   * @param masterAddress the master address
-   */
-  public FileSystemMasterClient(Subject subject, InetSocketAddress masterAddress) {
-    super(subject, masterAddress);
-  }
-
-  /**
-   * Creates a new file system master client.
-   *
-   * @param masterAddress the master address
-   */
-  public FileSystemMasterClient(InetSocketAddress masterAddress) {
-    super(null, masterAddress);
-  }
-
-  @Override
-  protected AlluxioService.Client getClient() {
-    return mClient;
-  }
-
-  @Override
-  protected String getServiceName() {
-    return Constants.FILE_SYSTEM_MASTER_CLIENT_SERVICE_NAME;
-  }
-
-  @Override
-  protected long getServiceVersion() {
-    return Constants.FILE_SYSTEM_MASTER_CLIENT_SERVICE_VERSION;
-  }
-
-  @Override
-  protected void afterConnect() throws IOException {
-    mClient = new FileSystemMasterClientService.Client(mProtocol);
-  }
+public interface FileSystemMasterClient extends MasterClient {
 
   /**
    * Checks the consistency of Alluxio metadata against the under storage for all files and
@@ -99,21 +43,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws AlluxioException if an Alluxio error occurs
    * @throws IOException if an I/O error occurs
    */
-  public synchronized List<AlluxioURI> checkConsistency(final AlluxioURI path,
-      final CheckConsistencyOptions options) throws AlluxioException, IOException {
-    return retryRPC(new RpcCallableThrowsAlluxioTException<List<AlluxioURI>>() {
-      @Override
-      public List<AlluxioURI> call() throws AlluxioTException, TException {
-        List<String> inconsistentPaths =
-            mClient.checkConsistency(path.getPath(), options.toThrift());
-        List<AlluxioURI> inconsistentUris = new ArrayList<>(inconsistentPaths.size());
-        for (String path : inconsistentPaths) {
-          inconsistentUris.add(new AlluxioURI(path));
-        }
-        return inconsistentUris;
-      }
-    });
-  }
+  List<AlluxioURI> checkConsistency(AlluxioURI path, CheckConsistencyOptions options)
+      throws AlluxioException, IOException;
 
   /**
    * Creates a new directory.
@@ -123,16 +54,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void createDirectory(final AlluxioURI path,
-      final CreateDirectoryOptions options) throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.createDirectory(path.getPath(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void createDirectory(AlluxioURI path, CreateDirectoryOptions options)
+      throws IOException, AlluxioException;
 
   /**
    * Creates a new file.
@@ -142,16 +65,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void createFile(final AlluxioURI path, final CreateFileOptions options)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.createFile(path.getPath(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void createFile(AlluxioURI path, CreateFileOptions options) throws IOException, AlluxioException;
 
   /**
    * Marks a file as completed.
@@ -161,16 +75,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void completeFile(final AlluxioURI path, final CompleteFileOptions options)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.completeFile(path.getPath(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void completeFile(AlluxioURI path, CompleteFileOptions options)
+      throws IOException, AlluxioException;
 
   /**
    * Deletes a file or a directory.
@@ -180,16 +86,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void delete(final AlluxioURI path, final DeleteOptions options)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.remove(path.getPath(), options.isRecursive());
-        return null;
-      }
-    });
-  }
+  void delete(AlluxioURI path, DeleteOptions options) throws IOException, AlluxioException;
 
   /**
    * Frees a file.
@@ -199,16 +96,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void free(final AlluxioURI path, final FreeOptions options)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.free(path.getPath(), options.isRecursive(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void free(AlluxioURI path, FreeOptions options) throws IOException, AlluxioException;
 
   /**
    * @param path the file path
@@ -216,15 +104,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized URIStatus getStatus(final AlluxioURI path) throws IOException,
-      AlluxioException {
-    return retryRPC(new RpcCallableThrowsAlluxioTException<URIStatus>() {
-      @Override
-      public URIStatus call() throws AlluxioTException, TException {
-        return new URIStatus(ThriftUtils.fromThrift(mClient.getStatus(path.getPath())));
-      }
-    });
-  }
+  URIStatus getStatus(AlluxioURI path) throws IOException, AlluxioException;
 
   /**
    * @param path the file path
@@ -232,15 +112,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized long getNewBlockIdForFile(final AlluxioURI path)
-      throws IOException, AlluxioException {
-    return retryRPC(new RpcCallableThrowsAlluxioTException<Long>() {
-      @Override
-      public Long call() throws AlluxioTException, TException {
-        return mClient.getNewBlockIdForFile(path.getPath());
-      }
-    });
-  }
+  long getNewBlockIdForFile(AlluxioURI path) throws IOException, AlluxioException;
 
   /**
    * @param path the path to list
@@ -249,20 +121,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized List<URIStatus> listStatus(final AlluxioURI path,
-      final ListStatusOptions options) throws IOException, AlluxioException {
-    return retryRPC(new RpcCallableThrowsAlluxioTException<List<URIStatus>>() {
-      @Override
-      public List<URIStatus> call() throws AlluxioTException, TException {
-        List<URIStatus> result = new ArrayList<URIStatus>();
-        for (alluxio.thrift.FileInfo fileInfo : mClient
-            .listStatus(path.getPath(), options.toThrift())) {
-          result.add(new URIStatus(ThriftUtils.fromThrift(fileInfo)));
-        }
-        return result;
-      }
-    });
-  }
+  List<URIStatus> listStatus(AlluxioURI path, ListStatusOptions options)
+      throws IOException, AlluxioException;
 
   /**
    * Loads the metadata of a file from the under file system.
@@ -274,15 +134,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @deprecated since version 1.1 and will be removed in version 2.0
    */
   @Deprecated
-  public synchronized void loadMetadata(final AlluxioURI path,
-      final LoadMetadataOptions options) throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Long>() {
-      @Override
-      public Long call() throws AlluxioTException, TException {
-        return mClient.loadMetadata(path.toString(), options.isRecursive());
-      }
-    });
-  }
+  void loadMetadata(AlluxioURI path, LoadMetadataOptions options)
+      throws IOException, AlluxioException;
 
   /**
    * Mounts the given UFS path under the given Alluxio path.
@@ -293,17 +146,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws AlluxioException if an Alluxio error occurs
    * @throws IOException an I/O error occurs
    */
-  public synchronized void mount(final AlluxioURI alluxioPath, final AlluxioURI ufsPath,
-      final MountOptions options)
-      throws AlluxioException, IOException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.mount(alluxioPath.toString(), ufsPath.toString(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void mount(AlluxioURI alluxioPath, AlluxioURI ufsPath, MountOptions options)
+      throws AlluxioException, IOException;
 
   /**
    * Renames a file or a directory.
@@ -313,16 +157,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void rename(final AlluxioURI src, final AlluxioURI dst)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.rename(src.getPath(), dst.getPath());
-        return null;
-      }
-    });
-  }
+  void rename(AlluxioURI src, AlluxioURI dst) throws IOException, AlluxioException;
 
   /**
    * Sets the file or directory attributes.
@@ -332,16 +167,8 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws AlluxioException if an Alluxio error occurs
    */
-  public synchronized void setAttribute(final AlluxioURI path, final SetAttributeOptions options)
-      throws IOException, AlluxioException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.setAttribute(path.getPath(), options.toThrift());
-        return null;
-      }
-    });
-  }
+  void setAttribute(AlluxioURI path, SetAttributeOptions options)
+      throws IOException, AlluxioException;
 
   /**
    * Schedules the async persistence of the given file.
@@ -350,16 +177,7 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws AlluxioException if an Alluxio error occurs
    * @throws IOException if an I/O error occurs
    */
-  public synchronized void scheduleAsyncPersist(final AlluxioURI path)
-      throws AlluxioException, IOException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.scheduleAsyncPersist(path.getPath());
-        return null;
-      }
-    });
-  }
+  void scheduleAsyncPersist(AlluxioURI path) throws AlluxioException, IOException;
 
   /**
    * Unmounts the given Alluxio path.
@@ -368,14 +186,5 @@ public final class FileSystemMasterClient extends AbstractMasterClient {
    * @throws AlluxioException if an Alluxio error occurs
    * @throws IOException an I/O error occurs
    */
-  public synchronized void unmount(final AlluxioURI alluxioPath)
-      throws AlluxioException, IOException {
-    retryRPC(new RpcCallableThrowsAlluxioTException<Void>() {
-      @Override
-      public Void call() throws AlluxioTException, TException {
-        mClient.unmount(alluxioPath.toString());
-        return null;
-      }
-    });
-  }
+  void unmount(AlluxioURI alluxioPath) throws AlluxioException, IOException;
 }
