+++ core/common/src/main/java/alluxio/AbstractClient.java
@@ -266,7 +266,7 @@ public abstract class AbstractClient implements Client {
   /**
    * @return the {@link InetSocketAddress} of the remote
    */
-  protected synchronized InetSocketAddress getAddress() {
+  public synchronized InetSocketAddress getAddress() {
     return mAddress;
   }
