+++ tests/src/test/java/alluxio/master/lineage/LineageMasterIntegrationTest.java
@@ -20,6 +20,7 @@ import alluxio.client.WriteType;
 import alluxio.client.file.FileOutStream;
 import alluxio.client.file.FileSystem;
 import alluxio.client.file.FileSystemMasterClient;
+import alluxio.client.file.RetryHandlingFileSystemMasterClient;
 import alluxio.client.file.URIStatus;
 import alluxio.client.file.options.CreateFileOptions;
 import alluxio.client.lineage.AlluxioLineage;
@@ -222,7 +223,7 @@ public class LineageMasterIntegrationTest {
   }
 
   private FileSystemMasterClient getFileSystemMasterClient() {
-    return new FileSystemMasterClient(null,
+    return new RetryHandlingFileSystemMasterClient(null,
         mLocalAlluxioClusterResource.get().getMaster().getAddress());
   }
 }
