+++ tests/src/test/java/alluxio/security/MasterClientAuthenticationIntegrationTest.java
@@ -15,6 +15,7 @@ import alluxio.AlluxioURI;
 import alluxio.LocalAlluxioClusterResource;
 import alluxio.PropertyKey;
 import alluxio.client.file.FileSystemMasterClient;
+import alluxio.client.file.RetryHandlingFileSystemMasterClient;
 import alluxio.client.file.options.CreateFileOptions;
 import alluxio.exception.ConnectionFailedException;
 import alluxio.security.authentication.AuthType;
@@ -87,7 +88,7 @@ public final class MasterClientAuthenticationIntegrationTest {
   public void customAuthenticationDenyConnect() throws Exception {
     mThrown.expect(ConnectionFailedException.class);
 
-    try (FileSystemMasterClient masterClient = new FileSystemMasterClient(
+    try (FileSystemMasterClient masterClient = new RetryHandlingFileSystemMasterClient(
         mLocalAlluxioClusterResource.get().getMaster().getAddress())) {
       Assert.assertFalse(masterClient.isConnected());
       // Using no-alluxio as loginUser to connect to Master, the IOException will be thrown
@@ -104,8 +105,8 @@ public final class MasterClientAuthenticationIntegrationTest {
    * @throws Exception if a {@link FileSystemMasterClient} operation fails
    */
   private void authenticationOperationTest(String filename) throws Exception {
-    FileSystemMasterClient masterClient =
-        new FileSystemMasterClient(mLocalAlluxioClusterResource.get().getMaster().getAddress());
+    FileSystemMasterClient masterClient = new RetryHandlingFileSystemMasterClient(
+        mLocalAlluxioClusterResource.get().getMaster().getAddress());
     Assert.assertFalse(masterClient.isConnected());
     masterClient.connect();
     Assert.assertTrue(masterClient.isConnected());
